Moving where CacheContainer is created and how it gets into Repository
