+++ src/GitHub.Api/Application/ApplicationManagerBase.cs
@@ -21,7 +21,6 @@ namespace GitHub.Unity
             UIScheduler = TaskScheduler.FromCurrentSynchronizationContext();
             ThreadingHelper.MainThreadScheduler = UIScheduler;
             TaskManager = new TaskManager(UIScheduler);
-            CacheContainer = new CacheContainer();
         }
 
         protected void Initialize()
@@ -117,7 +116,7 @@ namespace GitHub.Unity
                 .Then(GitClient.Commit("Initial commit", null))
                 .Then(_ =>
                 {
-                    Environment.InitializeRepository();
+                    Environment.InitializeRepository(CacheContainer);
                     RestartRepository();
                 })
                 .ThenInUI(InitializeUI);
@@ -130,7 +129,7 @@ namespace GitHub.Unity
             {
                 repositoryManager = Unity.RepositoryManager.CreateInstance(Platform, TaskManager, GitClient, Environment.RepositoryPath);
                 repositoryManager.Initialize();
-                Environment.Repository.Initialize(repositoryManager, CacheContainer);
+                Environment.Repository.Initialize(repositoryManager);
                 repositoryManager.Start();
                 Logger.Trace($"Got a repository? {Environment.Repository}");
             }
@@ -215,7 +214,7 @@ namespace GitHub.Unity
         public ISettings LocalSettings { get; protected set; }
         public ISettings SystemSettings { get; protected set; }
         public ISettings UserSettings { get; protected set; }
-        public ICacheContainer CacheContainer { get; private set; }
+        public ICacheContainer CacheContainer { get; protected set; }
         public IUsageTracker UsageTracker { get; protected set; }
 
         protected TaskScheduler UIScheduler { get; private set; }
